+++ BackEnd/src/main/java/com/manhattan/busyness_predictor/repository/HistoryRepository.java
@@ -11,17 +11,17 @@ import org.springframework.stereotype.Repository;
 import com.manhattan.busyness_predictor.model.History;
 
 @Repository
-public interface HistoryRepository extends JpaRepository<History, Long> {
+public interface HistoryRepository extends JpaRepository<History, Integer> {
 
-    List<History> findByUserId(Long userId);
+    List<History> findByUserId(Integer userId);
 
-    List<History> findByUserIdOrderByTimestampDesc(Long userId);
+    List<History> findByUserIdOrderByTimestampDesc(Integer userId);
 
-    List<History> findByLocationId(Long locationId);
+    List<History> findByLocationId(Integer locationId);
 
     // Find recent searches (last 30 days)
     @Query("SELECT h FROM History h WHERE h.userId = :userId AND h.timestamp > :since ORDER BY h.timestamp DESC")
-    List<History> findRecentSearches(@Param("userId") Long userId, @Param("since") LocalDateTime since);
+    List<History> findRecentSearches(@Param("userId") Integer userId, @Param("since") LocalDateTime since);
 
     // Get most searched locations
     @Query("SELECT h.locationId, COUNT(h) as searchCount FROM History h GROUP BY h.locationId ORDER BY searchCount DESC")
